+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 Development
 -----------
 
+- [Added] Implemented ``_query_gsm_for_lsm`` function to query the Global Sky Model 
+  database and return sources within a specified field of view as ``GlobalSkyModel`` objects.
 - Re-instated postgres backend storage starting from version 0.1.4.
   (`MR49 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/49>`__)
 - [Added] Watcher for creating local sky models from Flow entries